requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
boto3==1.20.26
botocore==1.23.26
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.7
django-allauth==0.44.0
django-cors-headers==3.10.0
django-dotenv==1.4.2
django-heroku==0.3.1
django-storages==1.12.3
django-versatileimagefield==2.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
djongo==1.3.6
dnspython==2.1.0
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==3.3
jmespath==0.10.0
oauthlib==3.1.1
Pillow==8.3.2
psycopg==3.0.7
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.3.0
pymongo==4.0.1
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.5
python-dotenv==0.19.2
python-magic==0.4.24
python3-openid==3.2.0
pytz==2021.3
razorpay==1.2.0
reportlab==3.6.2
requests==2.26.0
requests-oauthlib==1.3.0
s3transfer==0.5.0
six==1.16.0
sqlparse==0.2.4
twilio==7.1.0
urllib3==1.26.7
webencodings==0.5.1
whitenoise==5.3.0
xhtml2pdf==0.2.5
